changed datatypes

+++ src/upgradeInitializers/DiamondInit.sol
@@ -25,7 +25,7 @@ import {IERC721Receiver} from "../interfaces/IERC721Receiver.sol";
 contract DiamondInit {
     // You can add parameters to this function in order to pass in
     // data to set your own state variables
-    function init(uint256 ideationMarketFee) external {
+    function init(uint32 ideationMarketFee) external {
         // adding ERC165 data
         LibDiamond.DiamondStorage storage ds = LibDiamond.diamondStorage();
         ds.supportedInterfaces[type(IERC165).interfaceId] = true;
